Consolidate to single common L2Syncer
